requirements.txt
absl-py==2.1.0
astor==0.8.1
certifi==2022.12.7
click==8.1.8
colorama==0.4.6
cycler==0.11.0
Flask==2.2.5
fonttools==4.38.0
gast==0.7.0
grpcio==1.62.3
h5py==2.10.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.0.3
mock==5.2.0
numpy==1.16.6
packaging==24.0
Pillow==9.5.0
protobuf==3.20.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.1
scipy==1.7.3
six==1.17.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==2.3.0
typing_extensions==4.7.1
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
